fix test

+++ test/yukicoder/yuki_888.test.cpp
@@ -5,7 +5,7 @@ using namespace std;
 #include "../../lib/math/rho.hpp"
 
 int main(){
-    int n; cin >> n;
+    long long n; cin >> n;
     long long ans = 0;
     for(auto x : Rho::divisor(n)){
         ans += x;
